Made FileOpenPicker more generic
FileOpenPicker only worked with MainPage and the class that implemented
IFileOpenPickerContinuable had to be a Page.

Did work to make it more generic so any class that implements an
IFileOpen interface found in ContinuationManager can save a reference to
themselves to App.xaml.cs and be used during app reactivation when
arriving back from the Picker.