Added Helios.Controls local NuGet package
Managed to load my RadialRange control as a separate DLL using a local
NuGet package. Missing MVVM Light NuGet stuff was automatically added.

Still need to get it to build properly - right now I'm copying the DLLs
every time.
Also, the tick marks are broken, but that might be okay.

NOTE: the rest of the UI besides the RadialRange is collapsed.