Isolated the cause of the XAMLParseException
Isolated the cause of the XAMLParseException to RadialScale.RadialType

When I first started trying to port the Silverlight Gauge control to
XAML/Store App, I kept running into this undebuggable error in the XAML:
An exception of type 'Windows.UI.Xaml.Markup.XamlParseException'
occurred in Helios.exe but was not handled in user code

WinRT information: Cannot deserialize XBF metadata property list as '%1'
was not found in type '%0'. [Line: 0 Position: 0]

Additional information: The text associated with this error code could
not be found.

Learned how to port controls from this experience.
* Don't try to blindly copy/paste the code into the new Framework and
modify usings/namespaces appropriately - this won't work
* Instead, start with a blank TemplateControl
* Start adding in the code little by little
* Start with the simplest control mentioned in Generic.xaml - in this
project, that was MarkerIndicator
* Get it to draw on the screen
* Slowly start adding in only the necessary code
** Replace any dependency properties with static values
** In this project, that meant ignoring the Ranges, Ticks, and Labels
(in order of necessity) and setting all the Mins/Maxs to numeric values
* Once you get the basic control to draw, start switching the static
values back over to DependencyProperties and begin adding the
decorations back to the control (Ranges, Labels, Ticks)