ReorderListBox kind of works
Tried using Windows' GridView drag and reorder properties, but they
don't work. Instead, I ended up trying to port a Win Phone 7.1 control.

So far, lots of bugs but you can reorder an item. Next is to fix
scrolling.

Addendum to "Learned how to port"
* It's worth it to spend time before you start delving into the code to
see if the technologies you are attempting to use can feasibly be ported
* This involves a blind copy/paste, but then spending time to see the
overlap of the API surface of the framework porting from to framework
porting to
* Project probably builds but doesn't work at this point
* THEN follow my guidance from before (slow copy/paste effort after you
get it to build)