Bug fix on dragging
Fixed a bug introduced in last commit. Moved the logic to update the
items on either side of the dragged item out of the reoccuring part of
the ManipulationDelta method (second half) to the part where the image
is rendered (first half) - this makes it only happen once, which I think
is the right behavior.